transcriber/requirements.txt
openai-whisper
torch
tqdm
pyinstaller